Hazardous Commit: Split ContentNode from PathNode.
This is a massive commit that probably should have been on a branch if I was collaborating with anyone else.  But, since I'm not, whatevs.

The basic idea is that "ContentNode" has all the stuff for managing content, and "PathNode" has all the stuff for dealing with files or urls.  Some util functions are moved out into the global scope where they belong.

Right now, the build isn't working.